requirements.txt
minimalmodbus==2.1.1
paho-mqtt==1.6.1
pyserial==3.5
RPi.GPIO==0.7.1
